+++ GameSources/GoalObject.cpp
@@ -105,7 +105,7 @@ namespace basecross {
 
 		auto transComp = GetComponent<Transform>();
 		transComp->SetPosition(m_Position);
-		transComp->SetScale(m_Scale.x + 2, m_Scale.y + 7.5, m_Scale.z / (float)3);
+		transComp->SetScale(float(m_Scale.x + 2),float( m_Scale.y + 7.5), float(m_Scale.z / 3));
 		transComp->SetRotation(0.0f, XM_PIDIV2, 0.0f);
 
 		// ゴールオブジェクトの上の板ポリの追加
